fix(mesh): bug related to degenerated triangle when using smart linking
Reference array is not initialized properly for a degenerated triangle when using smart linking
Smart linking can merge two very close triangle vertices together. For instance a triangle T(a,b,c)
could become T(a,a,c) after smart linking. The UpdateReferences() function do not update the
refs[] array correctly in this case because it overwrites the same ref cell twice and leaves the
second ref cell not initialized because the indexer is updated only at the end of the function.